+++ BrawlLib/System/Windows/Controls/SCN0FogEditControl.cs
@@ -204,7 +204,7 @@ namespace System.Windows.Forms
             if (index >= 0)
             {
                 FogAnimationFrame f = (FogAnimationFrame)listKeyframes.SelectedItem;
-                numFrame.Value = f.Index + 1;
+                numFrame.Value = (decimal)(f.Index + 1);
             }
         }
